FLOE-116: refactored resourceInputPanel tests
Refactored the resourceInputPanel initialization test. This change had originally been done as part of the FLOE-171 work, but since master had been updated with the latest infusion I was able to move this change back into this branch. Also refactored a bit more to make use of a grade for the tested component.