FLOE-194: Converted the arrow to an active style
The arrow that sits between the icon and dialog is now part of the active styling of the icon. This will keep it centred below the icon as the dialog shifts (do to collision with the window border) below it. This shifting isn't dynamic, but calculated at the time the dialog is displayed.